fix(transform): try to redefine stack axis using only unmasked products
If the timestream that is input to CollateProducts is stacked, then
attempt to redefine its stack axis so that it only contains products
consisting of inputs that are not masked in the telescope instance.

Also moves determination of `bt_rev` from `process` to `setup` since
it only depends on the telescope instance.