refactor(svdfilter): transpose definition of SVD matrices
BREAKING CHANGE: Before performing the SVD, add an extra transpose, such that the U matrix
is now a matrix of the msign-nprod singular vectors, and V^dagger is a matrix of the
frequency singular vectors. This will lead to easier integration of the SVD matrix into
the other transforms (pol-SVD, KL) to be performed on the m-modes.