refactor(containers): use a `VisContainer` as the base for all visibility data
This adds a new baseclass which pulls out all the common functionality
for visibility like data, particularly that which deals with whether
datasetss are stacked or not. This adds two new properties
`.prodstack` for getting a characteristic feed pair for a stack, and
`.is_stacked` which says whether a dataset is stacked or not.