feat(SingleTask): more flexible control of the name of output files
This deprecates the current `output_root` scheme for naming output
files, with and `output_name` property that allows for finer and more
flexible control of the file name. Python format strings can be used to
construct names using various internal variables such as the task or key
name, or the tag or output count.