feat, docs (mockcatalog): add option for trivial selection function
If not bcat_path is passed to SelFuncEstimatorFromParams, it now uses a selection function that is unity everywhere.
Also in this commit, I updated a few lines in the sample config file in the header to work with the current
routines, and moved SelFuncEstimator below SelFuncEstiatorFromParams, because things weren't running if it was above.