refactor(ringmapmaker): refactor DeconvolveHybridM
- Move most of the contents of DeconvolveHybridM into
DeconvolveHybridMBase.

- Create subclasses that enable the user to deconvolve either the
driftscan analytical beam model or an external beam model.

- Create subclasses that enable a Tikhonov scheme for deconvolution
that uses a single regularisation parameter, or Weiner scheme that
assumes some model for the m-mode power spectrum of the signal.

- DeconvolveHybridMBase now computes the el-dependent EW dirty beam
and saves it to the output RingMap container if requested.

- DeconvolveHybridMBase now normalizes by the el-dependent
peak of the dirty beam so that point source peak flux is
preserved through the map making process.

- Create beam.py module and CreateBeamStream task that takes a
CommonModeGridBeam container and converts it to a HybridVisStream
container that can be m-mode transformed and used in the
deconvolving ringmapmakers.

- Add an el axis to the weight dataset in the RingMap container,
since the noise after deconvolution will be el dependent.

- Propagate noise variance through the deconvolution process.
Create RADependentWeight task that re-introduces the RA dependence
to the weights that was lost in the round-trip m-mode transform.
Applying this procedure to Gaussian noise realizations recovers
the standard deviation of the noise at the <10% level.