feat(flagging): create FindBeamformedOutliers and MaskBeamformedOutliers tasks (#138)
- Mask beamformed visibilities whose magnitude deviates
significantly from our expectation for thermal noise.

- Create three new containers: RingMapMask, FormedBeamMask,
and FormedBeamHAMask.  Each has a single mask dataset of the
same shape as the weight dataset in the container they are
named after.

- Split the previous MaskBeamformedOutliers task into
FindBeamformedOutliers and MaskBeamformedOutliers tasks that
generate and apply the mask, respectively.  This is useful
for applying the same mask constructed from data to
noise realizations.

- Add capability to expand each masked pixel by some fixed
number of pixels to capture nearby contaminated pixels that are
below the threshold.

- Scale the weight dataset in the beamformer by a factor of 2 to
account for the fact that we are taking the real componenet of the
beamformed visibility.  This missing factor was identified by
beamforming on a Gaussian noise realization and then comparing the
variance in the formed beams to the expectation based on the
weight dataset.