fix(RingMapBeamForm): exclude sources outside ringmap
The previous version of this task had a bug where a source with low enough
dec that it is below the lower boundary of the ringmap would nonetheless
be associated with a formed beam, but the beam would be at a location
within the ringmap determined by a wrapping of the za axis. In this commit,
we identify such sources and ensure that their beams and weights are zero.