refactor: use a consistent scheme for matching catalog and data coordinate epochs
This refactors the code such that `HealpixBeamForm`, `RingMapBeamform`
and `RingMapStack2D` all use a common scheme for determining what the
catalog data/epoch.

Additionally this factors out common functionality between
`RingMapBeamform` and `RingMapStack2D` this is equivalent, which also
fixes the bug when a ringmap of not the full ZA range is used in
`RingMapStack2D`.