feat(MapEllMSum): add direct SHT of ringmap
Previously, MapEllMSum only took a healpix map as input, so that
a ringmap had to be reprojected into a healpix map first. Now,
thanks to SHT routines in the ducc library, this task can operate
directly on a ringmap, with more accurate results than were obtained
from the ringmap-to-healpix conversion.