perf(dayenu): speed up filter construction/application
* Reformat the map at each (beam, pol, el) into a contiguous
array before applying the filter.  This speeds up the filter
application by roughly a factor of 200.

* Switch to using scipy.linalg.pinv.  This speeds up the
filter construction by 80%, however filter construction is
< 10% of the total runtime.

* Previously the highpass delay filter method stacked the
unique filters along the last axis.  Switch to stacking them
along the first axis.

* Use the same output format for the delay and m-mode filter.
Specifically return indices that map the unique filters to
the respective sub-arrays.  Previously the m-mode filter
created a copy of the filter for each element of the input
flag array, which was memory intensive.