perf(sourcestack): simple speed ups to SourceStack
Each rank was unnecessarily calculating quantities
for all sources and then down selecting those
quantities to the set of sources on that rank.
These operations were dominating the run time.
Fixing this resulted in an order of magnitude
improvement in run time, which is helpful when
stacking on a large number of mock catalogs.