fix(delay_spectrum_gibbs): fix typo in real/imag split of noise covariance
The previous code version took the inverse noise variance for each
complex-valued measurement in frequency space, and assigned 2**-0.5
times this variance to the real and imaginary parts. However, for a
circularly-symmetric complex Gaussian random number, the variances
of the real and imaginary parts should each be 0.5 times the total
variance, implying that the separate inverse variances should be
2 times the total inverse variance.

The only impact of this error on previous output of the code is that
the noise level assumed by the Gibbs sampler was slightly wrong.