feat(containers): modify the way that draco.ContainerBase handles _axes.
This commit tries to add some flexibility to the _axes attribute in
ContainerBase. _axes can be defined as either a tuple or a dict, which
can contain additional specs for that specific axis in the container.

This does not change the behaviour of any existing properties. An
additional property, .axes_spec, is added to access the fully resolved
_axes dict in the same way as .dataset_spec.

Closes chime-experiment/ch_pipeline#146