feat(task): modify default behaviour for task compression argument.
The expected behaviour of the compression argument was to overwrite
compression args if a dict with those args was provided, to disable
compression if something which evaluates to False was provided, and use
the default parameters for the container otherwise. The behaviour
previously would instead disable compression if no args were provided.