fix(sensitivity): fix bug in exclude_intracyl sensitivity calculation
Fixed bug in ComputeSystemSensitivity when exclude_intracyl was set
to true.  The inter-cylinder baselines were not being identified
correctly when calculating the `measured` sensitivity dataset due
to the CHIME cylinder rotation.  This resulted in 30 MHz ripples as
a function of frequency when calculating the normalized sensitivity
(`measured / radiometer`) from which the RFI mask is derived.