feat(ringmapmaker): new windows and option to ignore ns dirty beam
This provides two new windows for north-south beamforming,
triangular and tukey.  The triangular window approximates
natural weighting for chime, but can be scaled to ensure
achromatic synthesized beam response.  The tukey window
tapers the north and south ends of the cylinder, with the
fraction of cylinder tapered being an adjustable parameter.
Also, adds a flag to BeamformNS that if set does forces the
task to not save the dirty_beam to the container.