feat(container): allow copy of non-filtered datasets
Currently when using the copy_datasets_filter method of a container,
datasets that do not contain an axis that is being filtered are
not copied over.  This gives the option to copy over those datasets
as well (and in full since they have no filter applied).  This updates
the SelectFreq task to do this by default.