ss/daily filter (#291)
* feat(dayenu): add a task that delay filters hybrid beamformed visibilities

* feat(dayenu): add option to save filter that was applied

* feat(dayenu): add option for multiple stopbands in delay filter

* feat(sidereal): support for rebinning HybridVisStream containers

* feat(sidereal): capability to stack multiple HybridVisStream

* feat(transform): task that iterates over sub-bands of an input container

* feat(ringmapmaker): new windows and option to ignore ns dirty beam

This provides two new windows for north-south beamforming,
triangular and tukey.  The triangular window approximates
natural weighting for chime, but can be scaled to ensure
achromatic synthesized beam response.  The tukey window
tapers the north and south ends of the cylinder, with the
fraction of cylinder tapered being an adjustable parameter.
Also, adds a flag to BeamformNS that if set does forces the
task to not save the dirty_beam to the container.

* feat(ringmapmaker): do not require dirty_beam in BeamformEW task

* fix(ringmapmaker): add missing factors in weight propagation for BeamformEW

* feat(ringmapmaker): option to flag arbitary ew baselines in BeamformEW

* feat(container): allow copy of non-filtered datasets

Currently when using the copy_datasets_filter method of a container,
datasets that do not contain an axis that is being filtered are
not copied over.  This gives the option to copy over those datasets
as well (and in full since they have no filter applied).  This updates
the SelectFreq task to do this by default.

* feat(beamform): beamform to a catalog of sources from HybridVisStream

* feat(beamform): fit beamformed visibilities versus hour angle

Linear fit to a model consisting of an additive offset and the primary beam.
Also only add redshift dataset to FormedBeam containers if the catalog
has redshifts.

* feat(flagging): update masking tasks to handle other types of containers

* feat(flagging): update RFIMaskChisqHighDelay to apply sumthreshold

* feat(flagging): polarisation dependent masking

Capability to generate a polarisation dependent mask in
RFIMaskChisqHighDelay.  Capability to apply a polarisation
dependent mask with the ApplyTimeFreqMask task.

* feat(rfi): option to flag only positive excursions in sumthreshold