feat(transform): update MixData to enable jackknifes (#309)
* feat(transform): update MixData to enable jackknifes

MixData contains much of the functionality needed to perform
a jackknife of two datasets.  This adds two config flags that
(1) will invert the weights prior to mixing so that
uncertainty is propagated correctly and (2) flags any samples
where the weight is zero in any of the input data.  It also
creates a new Jackknife task that has the config properties
set to the correct values for performing a jackknife of two
datasets.