fix(dayenu): do not apply atten_threshold to fully masked samples (#315)
* fix(dayenu): do not apply atten_threshold to fully masked samples

Also ensure there are no masked frequencies in the data that were
unmasked during filter generation.

* fix(dayenu): preserve input container when apply_filter is False

Removes the line in DayenuDelayFilterHybridVis that enforces
identical masks for both polarizations, as this is handled when
the filter is applied. Updates the logic to ensure the input
container remains unmodified if apply_filter is set to False.