feat(noise): iterative noise realizations for power spectrum covariance (#323)
* Creates synthesis.noise.MultipleGaussianNoiseDatasets, which iteratively
generates noise realizations, cycling through the weight datasets provided
in a pair of input containers.  This is meant to accommodate the power
spectrum analysis, where we need to generate many noise realizations for
both even and odd day stacks.

* Creates analysis.delay.DelaySpectrumWienerEstimatorUpdatePowerSpectrum
class, which is the same as DelaySpectrumWienerEstimator, except that
the power spectrum used in the Wiener filter can change with each call
to process rather than being defined at setup.

* Creates core.io.LoadFilesAndSelect task, which loads a collection of
files on setup and then selects one to return on process based on the
attributes of some input container.

* Fixes bug in AccumulateList, so that it returns None on process_finish
when using group_size.