feat(ringmapmaker): add option to skip deconvolution (#347)
Add the `skip_deconvolution` config property to `DeconvolveHybridMBase`.
If set to True, then the map maker will apply the weighting and beam
model in m-mode space, but will not attempt to deconvolve the beam
model.  In this case, the map normalization is set by the
`reference_declination` property.

This also adds config properties `copy_weight` and `copy_tag` to
`ApplyDelayFilterHybridVis`. If set to True, then these quantities will
be copied over from the container holding the filter.