feat: track frequency-frequency noise covariance through pipeline
Applying the Dayenu filter introduces off-diagonal terms in the
frequency-frequency noise covariance matrix.  Update the following
tasks to calculate this covariance matrix:

    - DayenuDelayFilterHybridVis
    - DayenuDelayFilterHybridVis
    - DelayFilterHyFoReSBandpassHybridVisClean

Also update the following tasks to propagate this covariance matrix
through the pipeline:

    - SiderealRebinner
    - SiderealStacker
    - RADependentWeights

The covariance matrix is saved as a new dataset named either
freq_cov or complex_freq_cov in the HybridVisStream or RingMap,
similar to how the filter itself is saved.