feat: track freq-freq covariance matrix (#356)
Applying the Dayenu filter introduces off-diagonal terms in the
frequency-frequency noise covariance matrix.  Update the following
tasks to calculate this covariance matrix:

    - DayenuDelayFilterHybridVis
    - DayenuDelayFilterHybridVis
    - DelayFilterHyFoReSBandpassHybridVisClean

Also update the following tasks to propagate this covariance matrix
through the pipeline:

    - SiderealRebinner
    - SiderealStacker
    - RADependentWeights
    - SelectPol

The covariance matrix is saved as a new dataset named either
freq_cov or complex_freq_cov in the HybridVisStream or RingMap,
similar to how the filter itself is saved.

Creates two new tasks:

    - ConstructWienerDelayTransform
    - ApplyWienerDelayTransform

that implement a Wiener-filtered delay transform that:
    - handles frequency masks with non-uniform coverage
    - accounts for filtering applied to the data
    - uses the tracked frequency-frequency noise covariance
    - estimates the signal covariance from simulations

This approach avoids the need to estimate the delay power spectrum
via Gibbs+Wiener or NRML methods, instead assuming 21 cm signal + noise
and relying on the propagated frequency-domain covariances.

Creates a ReconstructVisFreqCov task that generates a FreqNoiseModel
container with the ingredients to synthesize noise with a specified
frequency-frequency covariance. It reuses logic from ReconstructVisWeight,
which is now refactored into a shared base class, ReconstructVisNoiseBase.

Creates FreqCorrelatedNoise and MultipleFreqCorrelatedNoise tasks that
use the FreqNoiseModel container to synthesize noise with the correct
correlations in frequency and el.

Finally, creates a TaperDelayTransform task that can apply a taper to
the delay transform. The existing task will not work for this purpose due to
the way the axes are collaped in the DelayTransform container.