fix(powerspec): remove window from covariance in wiener filter
Previous version of the code yielded identical results for different
window functions.  I think this is because the window was being
included in the forward modeling.  This resulted in the squared window
being applied to both the signal and noise covariance, which cancelled
out the factor that was being applied to the data.  Change the code to
only apply a single factor of the window to the data.

This commit also updates how the delay axis is set: if part of the band
is masked via window_lower_freq and window_upper_freq, only the unmasked
portion is now used to compute the minimum delay.