feat(powerspec): use inverse instead of pseudo-inverse (#362)
Calculate the inverse of the covariance matrix after removing
unmasked frequencies, rather than calculating the pseudo-inverse.
Use Cholesky factorization when calculating the inverse, since the
covariance matrix is Hermitian and positive definite.

This also  remove the window from the covariances in the wiener filter.
The previous version of the code yielded identical results for different
window functions.  I think this is because the window was being
included in the forward modeling.  This resulted in the squared window
being applied to both the signal and noise covariance, which cancelled
out the factor that was being applied to the data.  Change the code to
only apply a single factor of the window to the data.

This commit also updates how the delay axis is set: if part of the band
is masked via window_lower_freq and window_upper_freq, only the unmasked
portion is now used to compute the minimum delay.